Rename the dev script folders
