+++ packages/ui/dev/node-mocks/commands/vote.ts
