+++ packages/ui/dev/node-mocks/lib/isStage.ts
