+++ packages/ui/dev/query-node-mocks/generators/forum/generateCategories.ts
