+++ packages/ui/dev/query-node-mocks/generators/forum/generateForumMocks.ts
