+++ packages/ui/dev/query-node-mocks/generators/generateOpeningsAndUpcomingOpenings.ts
