+++ packages/ui/dev/query-node-mocks/generators/generateProposals.ts
