+++ packages/ui/dev/query-node-mocks/generators/generateWorkingGroups.ts
