+++ packages/ui/package.json
@@ -4,19 +4,17 @@
   "license": "GPL-3.0-only",
   "scripts": {
     "build": "node --max_old_space_size=4096 ./node_modules/.bin/webpack --mode production --progress",
-    "queries:patch": "ts-node dev/scripts/patchSchema.ts",
-    "queries:generate": "graphql-codegen --config codegen.config.yml && yarn lint:fix",
-    "queries": "yarn queries:patch && yarn queries:generate",
-    "lint": "yarn lint:prettier --check && yarn lint:eslint",
-    "lint:fix": "yarn lint:eslint --fix && yarn lint:prettier --write --loglevel warn",
-    "lint:eslint": "eslint './{src,test}/**/*.{ts,tsx}'",
-    "lint:prettier": "yarn prettier './{src,test}/**/*.{ts,tsx,html}'",
-    "start": "NODE_OPTIONS=--max_old_space_size=8192 webpack serve --mode development --progress",
-    "test": "jest",
-    "storybook": "start-storybook -p 6006 -s public --no-manager-cache",
     "build:storybook": "build-storybook -o build-storybook",
-    "generate-mocks": "ts-node dev/scripts/generateMocks.ts",
-    "node-mocks": "ts-node dev/mocks/index.ts"
+    "lint": "yarn lint:prettier --check && yarn lint:eslint",
+    "lint:eslint": "eslint './{src,test}/**/*.{ts,tsx}'",
+    "lint:fix": "yarn lint:eslint --fix && yarn lint:prettier --write --loglevel warn",
+    "lint:prettier": "yarn prettier './{src,test}/**/*.{ts,tsx,html}'",
+    "node-mocks": "ts-node dev/node/index.ts",
+    "queries:generate": "graphql-codegen --config codegen.config.yml && yarn lint:fix",
+    "query-node-mocks": "ts-node dev/node-mocks/generateMocks.ts",
+    "start": "NODE_OPTIONS=--max_old_space_size=8192 webpack serve --mode development --progress",
+    "storybook": "start-storybook -p 6006 -s public --no-manager-cache",
+    "test": "jest"
   },
   "dependencies": {
     "@apollo/client": "^3.3.21",
